Update README.md
Updated CFW FAQ with the latest details. Added some additional references to Tadpole, and added credits for `jasongrieves_02643` who has been doing a lot of recent development work on the tool lately. Updated the CPU section to reference the fact that the CPU clock has been increased in the latest firmware. Added links to `iq_132`'s Bluetooth mod. Added a section about the `.zfb` files used in the arcade section. Rearranged the details in the Bootloader Bug section to hopefully make things a bit clearer.