Update README.md
Refactored the CFW FAQ to make the hcRTOS and multicore efforts more distinct. Refactored the shortcuts FAQ to mention Tadpole first (easier for most folks). Refactored the slow SNES FAQ in light of 1.71. Refactored the quiet A/V output FAQ to suggest firmware updating first. Added details about the new 1.71 firmware version to the Firmware and Resources sections, and did a lot of little edits throughout the doc as well to standardise on Data Frog's version numbering (seems to be what most folks refer to). Added a note that the permanent bootload bugfix only needs to be applied once per device. Added another modification example to the analysis of `Foldername.ini` courtesy of `wyverino`.