Update README.md
Updated the is-this-any-good FAQ to mention screen tearing (thanks for the suggestion `superbottle`!), as well as references to the demonstration of the screen tearing fix. Updated the CFW FAQ with the latest details regarding the multicore alpha. Added a note about the screen tearing fix demo to the Display section. Added a note to the Wireless Connectivity section to mention that wireless controllers compatible with the froggy cannot be used with non-froggy-style devices. Rearranged some files in the Used and Unused sections, and moved that one pesky unknown file to the Other section (thanks for the help, `bnister`!)