Updated README.md
Added a sub-section to Hardware about the USB-C port, along with a mention of the hardware hack to add USB-C to USB-C charging support (thanks for the suggestion, @uli42 - closes #12). Minor updates to the custom firmware FAQ. Updates to the Arcade section to note that the specific version of FBA on the SF2000 is a custom build with internal tweaks made by the firmware's developer. Made a minor correction to the .zfb Files section (had resolution dimensions swapped by mistake). Added a note to the Default ROMs section with the specific dimensions of the thumbnail image they contain. Added links to `amir16yp`'s corrected Hebrew menu translations (closes #11).